pyproject.toml
[project]
name = "smart-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "openai>=1.97.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "streamlit>=1.47.1",
]
